requirements.txt
Django>=4.0,<5.0
psycopg2>=2.8
gunicorn>=19.9.0,<20.1.0
django-debug-toolbar==4.3.0
django-import-export==4.0.0
django-simple-history==3.5.0
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
sqids==0.4.1
ulid-py==1.1.0